fix: repositioned the elements in the right pane to create a more intuitive and user-friendly layout
Changes:
Repositioned Merge Button: The "Merge to left" button has been moved to the left side of the header in the right pane, placing it closer to the center of the screen.
Repositioned Additions Count: The additions count information ("+4 additions") has been moved to the right side of the header.